This was an experiment before running across this PR
https://github.com/derekkraan/horde/pull/271

Long story short, its not a good idea to nest Horde Supervisors in this way, probably a better approach
is have the "sub" processes attached to the main HordeSupervisor (via the project Module) and maybe
monitor that process in the "parent". Or just rethink needing them to be parent/child processes.